Added history to Dataset and trained baseline
The dataset supports using a history of months
for the sample.
The baseline using a simple ViT with few layers
achieved an f1 score approx 0.03 higher
than a model that predicts its own input on the
test set. In other words, the baseline's
prediction is more accurate than assuming nothing
changes over the specified time interval.